Update all of the API to use helper functions
