+++ pymc3/plots/posteriorplot.py
@@ -84,7 +84,7 @@ def plot_posterior(trace, varnames=None, transform=identity_transform, figsize=N
                           alpha_level=alpha_level, ref_val=ref_val, rope=rope, **kwargs)
     else:
         if varnames is None:
-            varnames = get_default_varnames(trace, plot_transformed)
+            varnames = get_default_varnames(trace.varnames, plot_transformed)
 
         trace_dict = get_trace_dict(trace, varnames)
