+++ pymc3/tests/test_util.py
@@ -1,0 +1,29 @@
+import pytest
+
+from pymc3.distributions.transforms import Transform
+import pymc3 as pm
+
+
+class TestTransformName(object):
+    cases = [
+        ('var', 'var_test__'),
+        ('var_test_', 'var_test__test__')
+    ]
+    transform_name = 'test'
+
+    def test_get_transformed_name(self):
+        test_transform = Transform()
+        test_transform.name = self.transform_name
+        for name, transformed in self.cases:
+            assert pm.util.get_transformed_name(name, test_transform) == transformed
+
+    def test_is_transformed_name(self):
+        for name, transformed in self.cases:
+            assert pm.util.is_transformed_name(transformed)
+            assert not pm.util.is_transformed_name(name)
+
+    def test_get_untransformed_name(self):
+        for name, transformed in self.cases:
+            assert pm.util.get_untransformed_name(transformed) == name
+            with pytest.raises(ValueError):
+                pm.util.get_untransformed_name(name)
